+++ src/force-directed-graph/index.ts
@@ -1,2 +1,0 @@
-export * from './force-directed-graph.module';
-export * from './force-directed-graph.component';