+++ src/line-chart/index.ts
@@ -1,4 +1,0 @@
-export * from './line-chart.module';
-export * from './line-chart.component';
-export * from './line.component';
-export * from './line-series.component';